refactor: rename constant.ts to constants.ts
